Upgrade of the VM creation processus and ip attribution
A new attributs of proxmox stores now the current status of a VM creation (creating;, created and the potentials error). This attributes is now displayed when /vm/vmid is called, instead of the frontend guess of the vm status (it also brings a very important feature : the display of very accurate error messages)
It's also the end of we ensure there is an error by waiting for 5 of them to happen  to trust them. The backend is now very clear on it status and the frontend doesn't have to change of endpoint each second

The VM ip attribution job is fixed ! It was failing if ANY VM had an issue to set up his ip and wasn't applying a new ip for the other VM. It resists now to errors (with verbose !!)