Fix #21 The VM deletion is 100% rewritten with a heavy and resilient process able to 	- Delete a VM when an error occured during its creation 	- Delete a VM when the user wanted to do so (it was working before but without any feedback for the user) The issue was caused by the lack of app context when deleting the VM in the database
The process now 1. shutdown the VM and wait for it. 2. delete the VM and wait for it (in promxmox), 3. delete the VM from the db 4. Advertise the user on the current status, particulary when the VM is deleted by the user with a brand new style dealing with the suppression with feedback on the process. Not that the server deals with the deletion alone without any need for the user to keep its window open.